Interactividad Rama2 (#7)
* probando here react

* mapa react

* Map React

* Interactividad Zoom al mapa, apretar alt y mover el mapa para ver edificios

* ordenando el code de interactividad

* enter

* interacciones

* Interactividad Mapa2
