+++ public/index.html
@@ -1,64 +1,37 @@
 <!DOCTYPE html>
 <html lang="en">
-  <head>
-    <meta charset="utf-8" />
-    <link rel="icon" href="%PUBLIC_URL%/favicon.ico" />
-    <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <meta name="theme-color" content="#000000" />
-    <meta
-      name="description"
-      content="Web site created using create-react-app"
-    />
-    <link rel="apple-touch-icon" href="%PUBLIC_URL%/logo192.png" />
-    <!--
-      manifest.json provides metadata used when your web app is installed on a
-      user's mobile device or desktop. See https://developers.google.com/web/fundamentals/web-app-manifest/
-    -->
-    <link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
-    <!--
-      Notice the use of %PUBLIC_URL% in the tags above.
-      It will be replaced with the URL of the `public` folder during the build.
-      Only files inside the `public` folder can be referenced from the HTML.
 
-      Unlike "/favicon.ico" or "favicon.ico", "%PUBLIC_URL%/favicon.ico" will
-      work correctly both with client-side routing and a non-root public URL.
-      Learn how to configure a non-root public URL by running `npm run build`.
-    -->
-    <title>React App</title>
-    <meta name="viewport" content="initial-scale=1.0, width=device-width" />
-    <script
-      type="text/javascript"
-      src="https://js.api.here.com/v3/3.1/mapsjs-core.js"
-    ></script>
-    <script
-      type="text/javascript"
-      src="https://js.api.here.com/v3/3.1/mapsjs-service.js"
-    ></script>
-    <script 
-      src="https://js.api.here.com/v3/3.1/mapsjs-mapevents.js" 
-      type="text/javascript" charset="utf-8">
-    </script>
-    <script 
-      src="https://js.api.here.com/v3/3.1/mapsjs-clustering.js" 
-      type="text/javascript" charset="utf-8">
-    </script>
-      <script 
-      src="https://js.api.here.com/v3/3.1/mapsjs-ui.js" 
-      type="text/javascript" charset="utf-8">
-    </script>
-  </head>
-  <body>
-    <noscript>You need to enable JavaScript to run this app.</noscript>
-    <div id="root"></div>
-    <!--
-      This HTML file is a template.
-      If you open it directly in the browser, you will see an empty page.
 
-      You can add webfonts, meta tags, or analytics to this file.
-      The build step will place the bundled scripts into the <body> tag.
+<head>
+  <title>React App</title>
+  <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no"> <!-- Nos entrega las flechas en el mapa -->
+  <script type="text/javascript" src="https://js.api.here.com/v3/3.1/mapsjs-core.js"></script>
+  <script type="text/javascript" src="https://js.api.here.com/v3/3.1/mapsjs-service.js"></script>
+  <script type="text/javascript" src="https://js.api.here.com/v3/3.1/mapsjs-ui.js"></script><!-- Nos entrega las flechas en el mapa -->
+  <script type="text/javascript" src="https://js.api.here.com/v3/3.1/mapsjs-mapevents.js"></script>
+  <!-- 23-03-2020-->
+  <script src="https://unpkg.com/react@16/umd/react.development.js" crossorigin></script>
+  <script src="https://unpkg.com/react-dom@16/umd/react-dom.development.js" crossorigin></script>
+   
+<link rel="stylesheet" type="text/css" href="https://js.api.here.com/v3/3.0/mapsjs-ui.css?dp-version=1526040296" />
 
-      To begin the development, run `npm start` or `yarn start`.
-      To create a production bundle, use `npm run build` or `yarn build`.
-    -->
-  </body>
+</head>
+
+
+<body>
+  <div id="root"></div>
+  <!--
+  Este archivo HTML es una plantilla.
+      Si lo abre directamente en el navegador, verá una página vacía.
+
+      Puede agregar fuentes web, metaetiquetas o análisis a este archivo.
+      El paso de compilación colocará los scripts agrupados en la etiqueta <body>.
+
+      Para comenzar el desarrollo, ejecute `npm start` o` yarn start`.
+      Para crear un paquete de producción, use `npm run build` o` yarn build`.
+    -->    
+   
+ 
+</body>
+
 </html>
