+++ src/App.js
@@ -1,14 +1,12 @@
 import React from 'react';
 import './App.css';
-import {DisplayMapClass} from './DisplayMapClass';
+import {DisplayMapClass} from './components/DisplayMapClass';
+
 
 function App() {
   return (
   <DisplayMapClass />
   );
   }
+  
   export default App;
-
-
-
-
