--- src/DisplayMapClass.js
@@ -1,47 +1,0 @@
-// src/DisplayMapClass.js
-import * as React from 'react';
-
-export class DisplayMapClass extends React.Component {
-  mapRef = React.createRef();
-
-  state = {
-    // The map instance to use during cleanup
-    map: null
-  };
-
-  componentDidMount() {
-
-    const H = window.H;
-    const platform = new H.service.Platform({
-          apikey: "kgSdjqerT-AJ-VhQvIwifffAhb9AuVu7LK3gMkbIweA"
-    });
-
-    const defaultLayers = platform.createDefaultLayers();
-
-    // Create an instance of the map
-    const map = new H.Map(
-      this.mapRef.current,
-      defaultLayers.vector.normal.map,
-      {
-        // This map is centered over Europe
-        center: { lat: 50, lng: 5 },
-        zoom: 4,
-        pixelRatio: window.devicePixelRatio || 1
-      }
-    );
-
-    this.setState({ map });
-  }
-
-  componentWillUnmount() {
-    // Cleanup after the map to avoid memory leaks when this component exits the page
-    this.state.map.dispose();
-  }
-
-  render() {
-    return (
-      // Set a height on the map so it will display
-      <div ref={this.mapRef} style={{ height: "500px" }} />
-    );
-  }
-}
