+++ src/setupTests.js
@@ -1,5 +1,5 @@
-// jest-dom adds custom jest matchers for asserting on DOM nodes.
-// allows you to do things like:
-// expect(element).toHaveTextContent(/react/i)
-// learn more: https://github.com/testing-library/jest-dom
+// jest-dom agrega matizadores de bromas personalizados para afirmar en nodos DOM.
+// te permite hacer cosas como:
+// esperar (elemento) .toHaveTextContent (/ react / i)
+// más información: https://github.com/testing-library/jest-dom
 import '@testing-library/jest-dom/extend-expect';
